All simcore warnings fixed. In one case (line 555 of nrn_setup.cpp it was a bug). Unused args are generally avoided by the statement (void)arg; Shadow warnings by changing names.

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@31d8161d4904d878f20e4b6e83580c699d1a8c4b

+++ src/simcore/nrniv/ivlistimpl.cpp
@@ -40,7 +40,7 @@ static long ListImpl_best_new_sizes[] = {
 };
 
 long ListImpl_best_new_count(long count, unsigned int size, unsigned int m) {
-    for (int i = 0; i < sizeof(ListImpl_best_new_sizes)/sizeof(long); i++) {
+    for (size_t i = 0; i < sizeof(ListImpl_best_new_sizes)/sizeof(long); i++) {
         if (count * size < ListImpl_best_new_sizes[i]) {
             return ListImpl_best_new_sizes[i] / size;
         }
